+++ sema4ai/vscode-client/src/actionServer.ts
@@ -277,19 +277,23 @@ const startActionServerInternal = async (directory: Uri) => {
     actionServerTerminal.sendText(`${location} start --port=${ACTION_SERVER_DEFAULT_PORT}`);
 };
 
+export const findActionPackagePath = async (): Promise<Uri | undefined> => {
+    // Need to list the action packages available to decide
+    // which one to use for the action server.
+    const selected = await listAndAskRobotSelection(
+        "Please select the Action Package from which the Action Server should load actions.",
+        "Unable to start Action Server because no Action Package was found in the workspace.",
+        { showActionPackages: true, showTaskPackages: false }
+    );
+    if (!selected) {
+        return;
+    }
+    return Uri.file(selected.directory);
+};
+
 export const startActionServer = async (directory: Uri) => {
     if (!directory) {
-        // Need to list the action packages available to decide
-        // which one to use for the action server.
-        const selected = await listAndAskRobotSelection(
-            "Please select the Action Package from which the Action Server should load actions.",
-            "Unable to start Action Server because no Action Package was found in the workspace.",
-            { showActionPackages: true, showTaskPackages: false }
-        );
-        if (!selected) {
-            return;
-        }
-        directory = Uri.file(selected.directory);
+        directory = await findActionPackagePath();
     }
 
     let actionServerTerminal: Terminal = getActionServerTerminal();